Work with float values instead of isize types to get better results and allow float values as input, improve some parts to get better precision

+++ src/functions.rs
@@ -1,15 +1,15 @@
-pub fn sum(f1: isize, f2: isize) -> isize {
+pub fn sum(f1: f64, f2: f64) -> f64 {
     f1 + f2
 }
 
-pub fn substraction(f1: isize, f2: isize) -> isize {
+pub fn substraction(f1: f64, f2: f64) -> f64 {
     f1 - f2
 }
 
-pub fn multiplication(f1: isize, f2: isize) -> isize {
+pub fn multiplication(f1: f64, f2: f64) -> f64 {
     f1 * f2
 }
 
-pub fn division(f1: isize, f2: isize) -> isize {
+pub fn division(f1: f64, f2: f64) -> f64 {
     f1 / f2
 }
